fix(pr): Add KeelReadOnlyRepository for plugin usage
