+++ keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/plugins/KeelServiceSdk.kt
@@ -1,12 +1,12 @@
-package com.netflix.spinnaker.keel.plugins
+package com.netflix.spinnaker.keel.api.plugins
 
 import com.netflix.spinnaker.keel.api.actuation.TaskLauncher
-import com.netflix.spinnaker.keel.persistence.KeelRepository
+import com.netflix.spinnaker.keel.api.persistence.KeelReadOnlyRepository
 
 /**
  * A simple SDK that can be consumed by external plugins to access core Keel functionality.
  */
 interface KeelServiceSdk {
-  val repository: KeelRepository
+  val repository: KeelReadOnlyRepository
   val taskLauncher: TaskLauncher
 }
