+++ keel-web/src/main/kotlin/com/netflix/spinnaker/keel/plugins/KeelServiceSdkImpl.kt
@@ -1,9 +1,10 @@
 package com.netflix.spinnaker.keel.plugins
 
 import com.netflix.spinnaker.keel.api.actuation.TaskLauncher
-import com.netflix.spinnaker.keel.persistence.KeelRepository
+import com.netflix.spinnaker.keel.api.persistence.KeelReadOnlyRepository
+import com.netflix.spinnaker.keel.api.plugins.KeelServiceSdk
 
 class KeelServiceSdkImpl(
-  override val repository: KeelRepository,
+  override val repository: KeelReadOnlyRepository,
   override val taskLauncher: TaskLauncher
 ) : KeelServiceSdk
